Seperated GUI related code to own file
